progress

+++ components/consumers/defectdojo/types/types.go
@@ -13,7 +13,7 @@ package types
 // FindingCreateRequest models all the possible fields in a POST /findings request.
 type FindingCreateRequest struct {
 	Active            bool     `json:"active"`
-	Cvssv3Score       float32  `json:"cvssv3_score,omitempty"`
+	Cvssv3Score       float64  `json:"cvssv3_score,omitempty"`
 	Cwe               int32    `json:"cwe,omitempty"`
 	Date              string   `json:"date,omitempty"`
 	Description       string   `json:"description"`
